🚚[Rename] 에러 파일들 컴포넌트로 분리
